Stage 2: complete the pointer-cast sweep to intptr_t
Step 5 missed ~60 pointer-to-int cast sites whose operands didn't
match original grep's pointer-name heuristics. The struct-pack
audit in Step 6 surfaced them — kv6 allocation code uses a
malloc-trailing-data layout where kv6->vox, kv6->xlen, kv6->ylen
are computed via pointer arithmetic on (int32_t)kv6 + sizeof(kv6data),
and the voxel-slab code stores v/ov/nv/v0/v1 pointer values into
vfifo[] integer arrays.

Flipped to intptr_t:
  - malloc allocation arithmetic: nkv6, newkv6, kv6, kv->vox,
    kv->xlen, spr->voxnum
  - slab pointers: v0, v1, ov, nv (with trailing-paren disambiguation
    for v1 vs value casts of v[1])
  - opticast zbuffer paths: p inside ((int32_t)p)+zbufoff
  - internal engine pointers: font6x8, gstartv, c->i1, pngdat,
    umost, dmost, pathashead, dpqdistlut, uurendmem, radarmem
  - (int32_t)malloc(...) result casts in voxlap5.c:5097/11752/11854/
    12396 and kplib.c:3395, and ((int32_t)buf)+ arithmetic in
    kplib.c:2424/2425
  - stray (int32_t)s on voxlap5.c:824 and (int32_t)v in
    vfifo[i+3] = ... on line 10704

Kept as (int32_t) — these are legitimate value widenings, not
pointer casts:
  - (int32_t)p->x where p is a dpoint3d* and ->x is a float/double
  - (int32_t)v[k] / (int32_t)tbuf[k] / (int32_t)slst[i].z[k] —
    byte-array subscript derefs being widened
  - (int32_t)SSWAPIB(...) in kplib's PNG/TGA/BMP header parsing
  - sizeof(int32_t), sizeof(olinbuf) expressions
  - function prototypes (grouscanasm, bswap, testflag)

Deliberately out of scope — deferred to the right stages:
  - kzopen's `return((int32_t)kzfs.fil)` pattern (public API — would
    need voxlap5.h signature change to intptr_t)
  - kpzload's `(*pic) = (intptr_t)malloc(...)` where *pic is
    int32_t — fixing the narrowing needs the int32_t* parameter in
    voxlap5.h to become intptr_t* (public API change)
  - Local `int32_t i = (intptr_t)malloc(j)` in voxlap5.c — the
    cast is honest now but narrows back at assignment; widening
    `i` itself is a type-propagation pass that's better done as
    prep for Stage 3's inline-asm rewrite
  - isnewfloatingadd/ot/chg signatures still take int32_t params

On 32-bit MSVC (LLP64) intptr_t and int32_t are both 4 bytes so
this is hash-neutral. The narrowing-at-boundaries issues above
are documented for Stage 3+ and don't affect 32-bit correctness.

+++ voxlap/kplib.c
@@ -2421,8 +2421,8 @@ static int32_t kddsrend (const char *buf, int32_t leng,
 	for(y=0;y<ysiz;y+=4,buf+=stride)
 		for(x=0;x<xsiz;x+=4)
 		{
-			if (dxt == 1) uptr = (unsigned char *)(((int32_t)buf)+(x<<1));
-						else uptr = (unsigned char *)(((int32_t)buf)+(x<<2)+8);
+			if (dxt == 1) uptr = (unsigned char *)(((intptr_t)buf)+(x<<1));
+						else uptr = (unsigned char *)(((intptr_t)buf)+(x<<2)+8);
 			c0 = SSWAPIB(*(unsigned short *)&uptr[0]);
 			r[0] = ((c0>>8)&0xf8); g[0] = ((c0>>3)&0xfc); b[0] = ((c0<<3)&0xfc); a[0] = 255;
 			c1 = SSWAPIB(*(unsigned short *)&uptr[2]);
@@ -3392,7 +3392,7 @@ VOXLAP_API extern void kpzload (const char *filnam, int32_t *pic, int32_t *bpl, 
 
 	kpgetdim(buf,leng,xsiz,ysiz);
 	(*bpl) = ((*xsiz)<<2);
-	(*pic) = (int32_t)malloc((*ysiz)*(*bpl)); if (!(*pic)) { free(buf); return; }
+	(*pic) = (intptr_t)malloc((*ysiz)*(*bpl)); if (!(*pic)) { free(buf); return; }
 	if (kprender(buf,leng,*pic,*bpl,*xsiz,*ysiz,0,0) < 0) { free(buf); free((void *)*pic); (*pic) = 0; return; }
 	free(buf);
 }
